pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gha-debug"
version = "0.1.0"
description = "Debug GitHub Actions workflows locally with step-by-step execution"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Intellirim"}]
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]
keywords = ["github-actions", "developer-tools", "cli", "ci-cd", "workflow-debugging", "devops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
gha-debug = "gha_debug.cli:main"

[project.urls]
Homepage = "https://github.com/intellirim/gha-debug"
